+++ client/plugos/syscalls/editor.ts
@@ -74,9 +74,7 @@ import { openSearchPanel } from "@codemirror/search";
 import { forceLinting } from "@codemirror/lint";
 import { refreshLintEffect } from "../../codemirror/lint.ts";
 import {
-  isValidPath,
-  parseToRef,
-  type Path,
+  coerceAndValidateRef,
   type Ref,
 } from "@silverbulletmd/silverbullet/lib/ref";
 import { insertNewlineContinueMarkup } from "@codemirror/lang-markdown";
@@ -149,111 +147,15 @@ export function editorSyscalls(client: Client): SysCallMapping {
       replaceState = false,
       newWindow = false,
     ) => {
-      if (typeof ref === "string") {
-        const parsedRef = parseToRef(ref);
-        if (!parsedRef) {
-          throw new Error(
-            "Unable to parse string provided to `editor.navigate` as ref",
-          );
-        }
-        ref = parsedRef;
-      }
-
-      if (
-        // @ts-expect-error: Legacy support
-        ref.page !== undefined
-      ) {
-        console.warn(
-          "You are using legacy navigation syntax with `editor.navigate()`, this will be phased out in the future",
-        );
-
-        const legacyRef = ref as unknown as {
-          kind: "page" | "document";
-          page: string;
-          pos?: number | { line: number; column: number };
-          header?: string;
-          meta?: boolean;
-        };
-
-        legacyRef.kind ??= "page";
-
-        let details: Ref["details"];
-
-        if (typeof legacyRef.pos === "number") {
-          details = {
-            type: "position",
-            pos: legacyRef.pos,
-          };
-        } else if (legacyRef.pos) {
-          details = {
-            type: "linecolumn",
-            line: legacyRef.pos.line,
-            column: legacyRef.pos.column,
-          };
-        } else if (legacyRef.header) {
-          details = {
-            type: "header",
-            header: legacyRef.header,
-          };
-        }
-
-        ref = {
-          path: (legacyRef.kind === "page"
-            ? `${legacyRef.page}.md`
-            : legacyRef.page) as Path,
-          details,
-          meta: legacyRef.meta,
-        };
-      }
-
-      // This validation code should be connected to the Ref type. Ideally using
-      // some validation library. Didn't want to use jsonschemas here tho.
-      // Ideally this would be moved into a function too
-      if (!isValidPath(ref.path) && ref.path !== "") {
-        throw new Error("Path passed in ref to `editor.navigate` is invalid");
-      } else if (typeof ref.meta !== "boolean" && ref.meta !== undefined) {
-        throw new Error("ref.meta has to be of type `boolean`");
-      } else if (ref.details !== undefined && typeof ref.details !== "object") {
-        throw new Error(
-          "ref.details has to be of type `object` or `undefined`",
-        );
-      } else if (
-        ref.details &&
-        !["position", "linecolumn", "header", "anchor"].includes(
-          ref.details.type,
-        )
-      ) {
-        throw new Error(
-          "ref.details.type has to be 'position', 'linecolumn', 'header' or 'anchor'",
-        );
-      }
-
-      if (
-        ref.details?.type === "position" &&
-        typeof ref.details.pos !== "number"
-      ) {
-        throw new Error("ref.details.pos has to be of type `number`");
-      } else if (
-        ref.details?.type === "header" &&
-        typeof ref.details.header !== "string"
-      ) {
-        throw new Error("ref.details.header has to be of type `string`");
-      } else if (
-        ref.details?.type === "linecolumn" &&
-        typeof ref.details.line !== "number" &&
-        typeof ref.details.column !== "number"
-      ) {
-        throw new Error(
-          "ref.details.line and ref.details.column has to be of type `number`",
-        );
-      } else if (
-        ref.details?.type === "anchor" &&
-        typeof ref.details.name !== "string"
-      ) {
-        throw new Error("ref.details.name has to be of type `string`");
-      }
-
-      await client.navigate(ref, replaceState, newWindow);
+      await client.navigate(coerceAndValidateRef(ref), replaceState, newWindow);
+    },
+    "editor.open": async (
+      _ctx,
+      ref: Ref | string,
+      replaceState = false,
+      newWindow = false,
+    ) => {
+      await client.open(coerceAndValidateRef(ref), replaceState, newWindow);
     },
     "editor.reloadPage": async () => {
       await client.reloadEditor();